requirements.txt
aiogram~=3.13.1
cloudscraper~=1.2.71
validators~=0.34.0
python-dotenv~=1.0.1
requests~=2.32.3
SQLAlchemy~=2.0.36
aiomysql
regex
